Split rust backend
Split into separate server and crawler implementation crates. Add
Cargo.toml to root directory to allow for testing in the stared `tests`
directory.
