+++ .gitignore
@@ -54,3 +54,8 @@ Thumbs.db
 # pixi environments
 .pixi/*
 !.pixi/config.toml
+
+
+# Added by cargo
+
+/target
