+++ Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "wikipedia-path-finder"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+tokio = { version = "1.38.0", features = ["full"] }
+crawler-rs = { path = "./crawler_rs/" }
