+++ crawler_rs/Cargo.toml
@@ -1,16 +1,10 @@
 [package]
-name = "api-rust"
+name = "crawler-rs"
 version = "0.1.0"
 edition = "2021"
 
 [dependencies]
 anyhow = "1.0.86"
-axum = "0.7.5"
-dotenvy = "0.15.7"
-http = "1.1.0"
 reqwest = "0.12.4"
 scraper = "0.19.0"
-serde = "1.0.203"
-serde_json = "1.0.117"
 tokio = { version = "1.38.0", features = ["full"] }
-tower-http = { version = "0.5.2", features = ["cors"] }
