+++ server_rs/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "server-rs"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+axum = "0.7.5"
+dotenvy = "0.15.7"
+http = "1.1.0"
+serde = "1.0.203"
+serde_json = "1.0.117"
+tokio = { version = "1.38.0", features = ["full"] }
+tower-http = { version = "0.5.2", features = ["cors"] }
+crawler-rs = { path = "../crawler_rs/" }
